#17 moved configuration to exported package to be usable from OSGi
