+++ java/src/main/java/org/eclipse/ditto/client/configuration/DummyAuthenticationConfiguration.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.ditto.client.configuration.internal;
+package org.eclipse.ditto.client.configuration;
 
 import static java.util.Objects.requireNonNull;
 import static org.eclipse.ditto.model.base.common.ConditionChecker.checkNotNull;
@@ -22,9 +22,6 @@ import java.util.Objects;
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
-
-import org.eclipse.ditto.client.configuration.AuthenticationConfiguration;
-import org.eclipse.ditto.client.configuration.ProxyConfiguration;
 
 /**
  * An implementation for {@link org.eclipse.ditto.client.configuration.AuthenticationConfiguration} based on "Ditto Dummy authentication" used for local
