--- java/src/main/java/org/eclipse/ditto/client/configuration/internal/package-info.java
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2019 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-
-/**
- * Internal configuration package of the Ditto Client which is not part of the public API.
- */
-@org.eclipse.ditto.utils.jsr305.annotations.AllParametersAndReturnValuesAreNonnullByDefault
-package org.eclipse.ditto.client.configuration.internal;
