+++ java/src/test/java/org/eclipse/ditto/client/DittoClientUsageExamples.java
@@ -37,10 +37,10 @@ import org.eclipse.ditto.client.changes.ChangeAction;
 import org.eclipse.ditto.client.configuration.MessagingConfiguration;
 import org.eclipse.ditto.client.configuration.ProxyConfiguration;
 import org.eclipse.ditto.client.configuration.TrustStoreConfiguration;
-import org.eclipse.ditto.client.configuration.internal.BasicAuthenticationConfiguration;
-import org.eclipse.ditto.client.configuration.internal.ClientCredentialsAuthenticationConfiguration;
-import org.eclipse.ditto.client.configuration.internal.DummyAuthenticationConfiguration;
-import org.eclipse.ditto.client.configuration.internal.WebSocketMessagingConfiguration;
+import org.eclipse.ditto.client.configuration.BasicAuthenticationConfiguration;
+import org.eclipse.ditto.client.configuration.ClientCredentialsAuthenticationConfiguration;
+import org.eclipse.ditto.client.configuration.DummyAuthenticationConfiguration;
+import org.eclipse.ditto.client.configuration.WebSocketMessagingConfiguration;
 import org.eclipse.ditto.client.messaging.AuthenticationProvider;
 import org.eclipse.ditto.client.messaging.AuthenticationProviders;
 import org.eclipse.ditto.client.messaging.MessagingProvider;
